% Generated by build script
\name{oalex_authors}
\alias{oalex_authors}
\title{Search OpenAlex for authors}
\usage{
oalex_authors()
}
\arguments{
\item{search}{Search query string (author name)}
\item{per_page}{Number of results. Default 10.}
}
\value{tibble: id, display_name, works_count, cited_by_count,}
\description{Search OpenAlex for authors}
