% Generated by build script
\name{oalex_institutions}
\alias{oalex_institutions}
\title{Search OpenAlex for institutions}
\usage{
oalex_institutions()
}
\arguments{
\item{search}{Search query string (institution name)}
\item{per_page}{Number of results. Default 10.}
}
\value{tibble: id, display_name, type, country_code, works_count,}
\description{Search OpenAlex for institutions}
